% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO_gen.R
\name{PO.sim}
\alias{PO.sim}
\title{generate data for the example}
\usage{
PO.sim(
  n,
  beta,
  Z.gen = function(n, p) matrix(rnorm(n * p), n, p),
  inv.m = exp,
  C.gen = function(n) runif(n, 0, 5)
)
}
\arguments{
\item{n}{Sample size}

\item{beta}{True parameter, p-dimensional vector}

\item{Z.gen}{Function to generate n by p covariate matrix}

\item{inv.m}{Inverse of baseline hazard function}

\item{C.gen}{Function to generate n-dimensional censoring time}
}
\description{
generate data for the example
}
