% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_PO_data}
\alias{sim_PO_data}
\title{Simulate data for PO function}
\format{
A list with 4 elements:
\describe{
\item{delta}{}
\item{X}{}
\item{Z}{}
\item{C}{}
}
}
\usage{
sim_PO_data
}
\description{
Simulate data for PO function
}
\keyword{datasets}
