% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO_gen.R
\name{PO.sim}
\alias{PO.sim}
\title{generate data for the example}
\usage{
PO.sim(
  n,
  beta,
  Z.gen = function(n, p) matrix(rnorm(n * p), n, p),
  inv.m = exp,
  C.gen = function(n) runif(n, 0, 5)
)
}
\arguments{
\item{n}{...}

\item{beta}{...}

\item{Z.gen}{...}

\item{inv.m}{...}

\item{C.gen}{...}
}
\description{
generate data for the example
}
